+++ tests/test-613vorbis_in_mp4.rb
@@ -1,0 +1,5 @@
+#!/usr/bin/ruby -w
+
+# T_613vorbis_in_mp4
+describe "mkvmerge / Vorbis in MP4"
+test_merge "data/mp4/vorbis.mp4", :args => "--no-video"